+++ app/users/models.py
@@ -1,0 +1,10 @@
+
+class User(object):
+    """docstring for ClassName"""
+
+    def __init__(self, email=None, password=None, role=None):
+        self.email = email
+        self.password = password
+        self.role = role
+
+
